Rename data and trainer modules
